Separate calibration file, ExecutedProcesses.js
- Calibration routines moved to a separate file
- Creating  ExecutedProcesses.js script file to run full processing
