+++ AutoIntegrateEngine.js
@@ -127,6 +127,7 @@ this.par = this.global.par;
 this.ppar = this.global.ppar;
 
 this.veralux = new AutoIntegrateVeraLuxHMS(this.global.debug);
+this.calibrate = new AutoIntegrateCalibrate(this.global, this.util, this.flowchart, this);
 
 
 // Copy of files from this.global object
@@ -140,14 +141,10 @@ this.standalone_narrowband_mappings = null;
 
 this.selectiveColor = null;
 
-this.executed_processes = [];
-
 this.ssweight_set = false;
 this.H_in_R_channel = false;
 this.is_luminance_images = false;    // Do we have luminance files from autocontinue or FITS
 this.autocontinue_prefix = "";       // prefix used to find base files for autocontinue
-this.creating_mask = false;          // flag for creating mask
-this.skip_process_value_save = false;   // flag for iterative stretching
 
 this.crop_truncate_amount = null;       // used when cropping channel images
 this.crop_lowClipImageName = null;      // integrated image used to calculate this.crop_truncate_amount
@@ -465,7 +462,7 @@ flowchartFilterFiles(fileNames, filetype)
       console.flush();
       for (var i = 0; i < allfilesarr.length; i++) {
             console.writeln("flowchartFilterFiles, " + allfilesarr[i].filter + ", " + allfilesarr[i].files.length);
-            var exptimeGroups = this.groupLightsByExposureTime(allfilesarr[i].files);
+            var exptimeGroups = this.calibrate.groupLightsByExposureTime(allfilesarr[i].files);
             for (var eg = 0; eg < exptimeGroups.length; eg++) {
                   for (var j = 0; j < exptimeGroups[eg].files.length; j++) {
                         var name = exptimeGroups[eg].files[j].name;
@@ -835,7 +832,7 @@ copy_processed_image(imgWin, process_name)
 {
       if (this.par.keep_processed_images.val 
           && imgWin 
-          && !this.creating_mask
+          && !this.global.creating_mask
           && this.global.is_processing == this.global.processing_state.processing) 
       {
             // Copy window to new window with process name
@@ -942,6 +939,7 @@ extractLchannel(sourceWindow, from_lights)
       var targetWindow = ImageWindow.activeWindow;
       sourceWindow.mainView.endProcess();
 
+      this.printAndSaveProcessValues(P, "", sourceWindow.mainView.id);
       this.engine_end_process(node);
       return targetWindow;
 }
@@ -1782,512 +1780,6 @@ writeImage(filePath, imageWindow)
       fileFormatInstance.close();
    
       return true;
-}
-
-/* 
- * Image calibration as described in Light Vortex Astronomy.
- */
-
-// Integrate (stack) bias and dark images
-runImageIntegrationBiasDarks(images, name, type, exptime)
-{
-      console.writeln("runImageIntegrationBiasDarks, images[0] " + images[0][1] + ", name " + name);
-      var exptimeTxt = (exptime != null && exptime > 0) ? " (" + exptime + "s)" : "";
-      var node = this.flowchart.flowchartOperation("ImageIntegration:" + type + exptimeTxt);
-
-      if (this.global.get_flowchart_data) {
-            return this.flowchartNewIntegrationImage(images[0][1], name);
-      }
-
-      this.ensureThreeImages(images);
-
-      var P = new ImageIntegration;
-      P.images = images; // [ enabled, path, drizzlePath, localNormalizationDataPath ];
-      P.rejection = this.getRejectionAlgorithm(images.length);
-      P.weightMode = ImageIntegration.DontCare;
-      P.normalization = ImageIntegration.NoNormalization;
-      P.rangeClipLow = false;
-      if (this.global.pixinsight_version_num < 1080812) {
-            P.evaluateNoise = false;
-      } else {
-            P.evaluateSNR = false;
-      }
-      P.subtractPedestals = false;
-
-      P.executeGlobal();
-
-      this.util.closeOneWindowById(P.highRejectionMapImageId);
-      this.util.closeOneWindowById(P.lowRejectionMapImageId);
-      this.util.closeOneWindowById(P.slopeMapImageId);
-
-      var new_name = this.util.windowRename(P.integrationImageId, name);
-
-      this.printAndSaveProcessValues(P, type);
-      this.engine_end_process(node);
-
-      this.setAutoIntegrateVersionIfNeeded(this.util.findWindow(new_name));
-
-      console.writeln("runImageIntegrationBiasDarks, integrated image " + new_name);
-
-      return new_name;
-}
-
-// Generate SuperBias from integrated bias image
-runSuberBias(biasWin)
-{
-      console.writeln("runSuberBias, bias " + biasWin.mainView.id);
-      var node = this.flowchart.flowchartOperation("Superbias");
-
-      if (this.global.get_flowchart_data) {
-            return this.flowchartNewImage(biasWin, this.ppar.win_prefix + "AutoMasterSuperBias");
-      }
-
-      var P = new Superbias;
-
-      biasWin.mainView.beginProcess(UndoFlag.NoSwapFile);
-
-      P.executeOn(biasWin.mainView, false);
-
-      biasWin.mainView.endProcess();
-
-      var targetWindow = ImageWindow.activeWindow;
-
-      this.util.windowRenameKeepif(targetWindow.mainView.id, this.ppar.win_prefix + "AutoMasterSuperBias", true);
-
-      this.printAndSaveProcessValues(P);
-      this.engine_end_process(node, biasWin, "Superbias", false);
-      
-      return targetWindow.mainView.id
-}
-
-/* Read FITS headers from a file and return its EXPTIME/EXPOSURE value. */
-getExptimeFromFile(filePath)
-{
-      var keywords = this.getFileKeywords(filePath);
-      for (var j = 0; j < keywords.length; j++) {
-            var value = keywords[j].strippedValue.trim();
-            if (keywords[j].name == "EXPTIME" || keywords[j].name == "EXPOSURE") {
-                  return parseFloat(value);
-            }
-      }
-      return 0;
-}
-
-/* Group dark file paths by their exposure time.
- * fileNames is an array of file path strings.
- * Returns array of { exptime: <number>, files: [<paths>] }.
- */
-groupDarksByExposureTime(fileNames)
-{
-      var groups = {};
-      for (var i = 0; i < fileNames.length; i++) {
-            var exptime = this.getExptimeFromFile(fileNames[i]);
-            var key = exptime.toString();
-            if (!groups[key]) {
-                  groups[key] = { exptime: exptime, files: [] };
-            }
-            groups[key].files.push(fileNames[i]);
-      }
-      var result = [];
-      for (var key in groups) {
-            result.push(groups[key]);
-      }
-      return result;
-}
-
-// Group light file-info objects (which already carry .exptime) by exposure time.
-// Returns array of { exptime, files[] } similar to groupDarksByExposureTime.
-groupLightsByExposureTime(filearr)
-{
-      var groups = {};
-      for (var i = 0; i < filearr.length; i++) {
-            var exptime = filearr[i].exptime;
-            var key = exptime.toString();
-            if (!groups[key]) {
-                  groups[key] = { exptime: exptime, files: [] };
-            }
-            groups[key].files.push(filearr[i]);
-      }
-      var result = [];
-      for (var key in groups) {
-            result.push(groups[key]);
-      }
-      return result;
-}
-
-/* Select the best matching master dark for a target exposure time.
- * masterdarkInfoArr may be a single path string, an array of { exptime, path }
- * objects, or null. Returns the best matching path, or null if none available.
- */
-selectMasterDarkForExptime(masterdarkInfoArr, targetExptime)
-{
-      if (masterdarkInfoArr == null) {
-            return null;
-      }
-      if (!Array.isArray(masterdarkInfoArr)) {
-            // Single master dark — warn if its exposure time differs too much from the lights.
-            var darkExptime = this.getExptimeFromFile(masterdarkInfoArr);
-            console.writeln("selectMasterDarkForExptime: single master dark " + masterdarkInfoArr + ", exptime " + darkExptime + "s, targetExptime " + targetExptime + "s");
-            if (darkExptime > 0 && Math.abs(darkExptime - targetExptime) > 0.1 * targetExptime) {
-                  this.util.addWarningStatus("Warning: Master dark exposure time " + darkExptime + "s does not closely match light exposure time " + targetExptime + "s");
-            }
-            return masterdarkInfoArr;
-      }
-      // Find exact or closest match
-      var bestIdx = 0;
-      var bestDiff = Math.abs(masterdarkInfoArr[0].exptime - targetExptime);
-      for (var i = 1; i < masterdarkInfoArr.length; i++) {
-            var diff = Math.abs(masterdarkInfoArr[i].exptime - targetExptime);
-            if (diff < bestDiff) {
-                  bestDiff = diff;
-                  bestIdx = i;
-            }
-      }
-      // Report a warning if exposure times differ too much (e.g. more than 10%)
-      var bestExptime = masterdarkInfoArr[bestIdx].exptime;
-      console.writeln("selectMasterDarkForExptime: best match " + masterdarkInfoArr[bestIdx].path + ", exptime " + bestExptime + "s, targetExptime " + targetExptime + "s");
-      if (bestDiff > 0.1 * targetExptime) {
-            this.util.addWarningStatus("Warning: No closely matching master dark found for exposure time " + targetExptime + ". Best match is " + bestExptime);
-      }
-      return masterdarkInfoArr[bestIdx].path;
-}
-
-/* Match a master file to images. There can be an array of
- * master files in which case we try find the best match.
- * Images is a list of [enabled, path]
- */
-matchMasterToImages(images, masterPath)
-{
-      if (!Array.isArray(masterPath)) {
-            console.writeln("matchMasterToImages, masterPath " + masterPath);
-            return masterPath;
-      }
-      if (masterPath.length == 1) {
-            console.writeln("matchMasterToImages, masterPath[0] " + masterPath[0]);
-            return masterPath[0];
-      }
-
-      /* Try find best match from masterPath for images.
-       * Pick first image file as a reference.
-       */
-      var imageWin = this.util.openImageWindowFromFile(images[0][1]);
-      console.writeln("matchMasterToImages, images[0][1] " + images[0][1]);
-      console.writeln("matchMasterToImages, imageWin.width " + imageWin.mainView.image.width + ", imageWin.height "+ imageWin.mainView.image.height);
-
-      /* Loop through master files and pick the matching one.
-       */
-      var matchingMaster = null;
-      for (var i = 0; i < masterPath.length; i++) {
-            var masterWin = this.util.openImageWindowFromFile(masterPath[i]);
-            console.writeln("matchMasterToImages, check masterPath[ " + i + "] " + masterPath[i]);
-            console.writeln("matchMasterToImages, masterWin.width " + masterWin.mainView.image.width + ", masterWin.height " + masterWin.mainView.image.height);
-            if (masterWin.mainView.image.width == imageWin.mainView.image.width 
-                && masterWin.mainView.image.height == imageWin.mainView.image.height)
-            {
-                  /* We have a match. */
-                  matchingMaster = masterPath[i];
-                  console.writeln("matchMasterToImages, found match " + matchingMaster);
-            }
-            this.util.closeOneWindow(masterWin);
-            if (matchingMaster != null) {
-                  break;
-            }
-      }
-      this.util.closeOneWindow(imageWin);
-
-      if (matchingMaster == null) {
-            this.util.throwFatalError("***matchMasterToImages Error: Can't find matching master file");
-      }
-
-      return matchingMaster;
-}
-
-// Run ImageCalibration to darks using master bias image
-// Output will be _c.xisf images
-runCalibrateDarks(fileNames, masterbiasPath)
-{
-      if (!this.par.bias_use_on_darks.val) {
-            console.noteln("runCalibrateDarks, bias use on darks is disabled");
-            return fileNames;
-      }
-      if (masterbiasPath == null) {
-            console.noteln("runCalibrateDarks, no master bias");
-            return fileNames;
-      }
-
-      console.noteln("runCalibrateDarks, images[0] " + fileNames[0][1] + ", master bias " + masterbiasPath);
-      console.writeln("runCalibrateDarks, master bias " + masterbiasPath);
-      var node = this.flowchart.flowchartOperation("ImageCalibration:darks");
-
-      if (this.global.get_flowchart_data) {
-            return fileNames;
-      }
-
-      var P = new ImageCalibration;
-      P.targetFrames = this.fileNamesToEnabledPath(fileNames); // [ enabled, path ];
-      P.outputPedestalMode = ImageCalibration.OutputPedestal_Auto;
-      P.enableCFA = this.is_color_files && this.local_debayer_pattern != 'None';
-      P.cfaPattern = this.global.debayerPattern_enums[this.global.debayerPattern_values.indexOf(this.local_debayer_pattern)];
-      P.masterBiasEnabled = true;
-      P.masterBiasPath = this.matchMasterToImages(P.targetFrames, masterbiasPath);
-      P.masterDarkEnabled = false;
-      P.masterFlatEnabled = false;
-      P.outputDirectory = this.global.outputRootDir + this.global.AutoOutputDir;
-      P.overwriteExistingFiles = true;
-
-      P.executeGlobal();
-
-      this.printAndSaveProcessValues(P, "darks");
-      this.engine_end_process(node);
-
-      return this.fileNamesFromOutputData(P.outputData);
-}
-
-// Run ImageCalibration to flats using master bias and master dark images
-// Output will be _c.xisf images
-runCalibrateFlats(images, masterbiasPath, masterdarkPath, masterflatdarkPath, filterName)
-{
-      if (!this.par.use_darks_on_flat_calibrate.val) {
-            masterdarkPath = null;
-      }
-      if (masterbiasPath == null && masterdarkPath == null && masterflatdarkPath == null) {
-            console.noteln("runCalibrateFlats, no master bias or dark");
-            return this.imagesEnabledPathToFileList(images);
-      }
-
-      var txt = "";
-      if (masterflatdarkPath != null) {
-            txt = this.appendTxtWithComma(txt, "flat darks");
-      } else if (masterbiasPath != null) {
-            txt = this.appendTxtWithComma(txt, "bias");
-      }
-      if (masterdarkPath != null && this.par.use_darks_on_flat_calibrate.val && masterflatdarkPath == null) {
-            txt = this.appendTxtWithComma(txt, "darks");
-      }
-      if (txt != "") {
-            txt = " (use " + txt + ")";
-      }
-
-      console.noteln("runCalibrateFlats, images[0] " + images[0][1]);
-      var node = this.flowchart.flowchartOperation("ImageCalibration:flats" + txt);
-
-      if (this.global.get_flowchart_data) {
-            return this.imagesEnabledPathToFileList(images);
-      }
-
-      var P = new ImageCalibration;
-      P.targetFrames = images; // [ // enabled, path ];
-      P.outputPedestalMode = ImageCalibration.OutputPedestal_Auto;
-      P.enableCFA = this.is_color_files && this.local_debayer_pattern != 'None';
-      P.cfaPattern = this.global.debayerPattern_enums[this.global.debayerPattern_values.indexOf(this.local_debayer_pattern)];
-      if (masterflatdarkPath != null) {
-            console.writeln("runCalibrateFlats, master flat dark " + masterflatdarkPath);
-            P.masterBiasEnabled = true;
-            P.masterBiasPath = this.matchMasterToImages(images, masterflatdarkPath);
-      } else if (masterbiasPath != null) {
-            console.writeln("runCalibrateFlats, master bias " + masterbiasPath);
-            P.masterBiasEnabled = true;
-            P.masterBiasPath = this.matchMasterToImages(images, masterbiasPath);
-      } else {
-            console.writeln("runCalibrateFlats, no master bias or flat dark");
-            P.masterBiasEnabled = false;
-            P.masterBiasPath = "";
-      }
-      if (masterdarkPath != null && this.par.use_darks_on_flat_calibrate.val && masterflatdarkPath == null) {
-            console.writeln("runCalibrateFlats, master dark " + masterdarkPath);
-            P.masterDarkEnabled = true;
-            P.masterDarkPath = this.matchMasterToImages(images, masterdarkPath);
-      } else {
-            console.writeln("runCalibrateFlats, no master dark");
-            P.masterDarkEnabled = false;
-            P.masterDarkPath = "";
-      }
-      P.masterFlatEnabled = false;
-      P.masterFlatPath = "";
-      P.calibrateBias = false;
-      if (this.darkIsBiasCalibrated(masterdarkPath)) {
-            P.calibrateDark = false;
-      } else {
-            P.calibrateDark = true;
-      }
-      P.outputDirectory = this.global.outputRootDir + this.global.AutoOutputDir;
-      P.overwriteExistingFiles = true;
-
-      P.executeGlobal();
-
-      this.printAndSaveProcessValues(P, "flats_" + filterName);
-      this.engine_end_process(node);
-
-      return this.fileNamesFromOutputData(P.outputData);
-}
-
-// Run image integration on flat frames
-// If you have stars in flat frames, set
-// P.pcClipLow and P.pcClipHigh to a 
-// tiny value like 0.010
-runImageIntegrationFlats(images, name, filterName)
-{
-      console.writeln("runImageIntegrationFlats, images[0] " + images[0][1] + ", name " + name);
-      var node = this.flowchart.flowchartOperation("ImageIntegration:flats");
-
-      if (this.global.get_flowchart_data) {
-            return this.flowchartNewIntegrationImage(images[0][1], name);
-      }
-
-      var P = new ImageIntegration;
-      // Flats must always use Average combination for a proper master flat;
-      // the user's integration_combination setting applies only to light frames.
-      P.combination = ImageIntegration.Average;
-            
-      P.images = images; // [ enabled, path, drizzlePath, localNormalizationDataPath ];
-      P.weightMode = ImageIntegration.DontCare;
-      P.normalization = ImageIntegration.MultiplicativeWithScaling;
-      P.rejection = ImageIntegration.PercentileClip;
-      P.rejectionNormalization = ImageIntegration.EqualizeFluxes;
-      if (this.par.stars_in_flats.val) {
-            P.pcClipLow = 0.010;
-            P.pcClipHigh = 0.010;
-      } else {
-            P.pcClipLow = 0.200;
-            P.pcClipHigh = 0.100;
-      }
-      P.rangeClipLow = false;
-      P.subtractPedestals = false;
-
-      P.executeGlobal();
-
-      this.util.closeOneWindowById(P.highRejectionMapImageId);
-      this.util.closeOneWindowById(P.lowRejectionMapImageId);
-      this.util.closeOneWindowById(P.slopeMapImageId);
-
-      var new_name = this.util.windowRename(P.integrationImageId, name);
-
-      this.printAndSaveProcessValues(P, "flats_" + filterName);
-      this.engine_end_process(node);
-
-      this.setAutoIntegrateVersionIfNeeded(this.util.findWindow(new_name));
-
-      console.writeln("runImageIntegrationFlats, integrated image " + new_name);
-
-      return new_name;
-}
-
-appendTxtWithComma(txt, append)
-{
-      if (txt != "") {
-            txt = txt + ", ";
-      }
-      return txt + append;
-}
-
-runCalibrateLights(images, masterbiasPath, masterdarkPath, masterflatPath, filterName, exptime)
-{
-      if (!this.par.bias_use_on_lights.val) {
-            if (masterdarkPath != null && !this.darkIsBiasCalibrated(masterdarkPath)) {
-                  // Dark still contains bias signal (not pre-calibrated), so subtracting
-                  // the dark from lights naturally cancels the bias — no need to apply
-                  // bias separately.
-                  console.noteln("runCalibrateLights, master dark has bias (CALSTAT has no B), bias cancels via dark subtraction, ignore master bias");
-                  masterbiasPath = null;
-            }
-            // When the dark had bias subtracted (CALSTAT has B, or pre_calibrate_darks
-            // is set), the dark has no bias, so master bias must be applied directly
-            // to lights — do not null masterbiasPath.
-      }
-      if (masterbiasPath == null && masterdarkPath == null && masterflatPath == null) {
-            console.noteln("runCalibrateLights, no master bias, dark or flat");
-            return this.imagesEnabledPathToFileList(images);
-      }
-
-      console.noteln("runCalibrateLights, images[0] " + images[0][1]);
-
-      var txt = "";
-      if (masterbiasPath != null) {
-            txt = this.appendTxtWithComma(txt, "bias");
-      }
-      if (masterdarkPath != null) {
-            txt = this.appendTxtWithComma(txt, "darks");
-      }
-      if (masterflatPath != null) {
-            txt = this.appendTxtWithComma(txt, "flats");
-      }
-      if (exptime != null && exptime > 0) {
-            txt = this.appendTxtWithComma(txt, exptime + "s");
-      }
-      if (txt != "") {
-            txt = " (" + txt + ")";
-      }
-
-      var node = this.flowchart.flowchartOperation("ImageCalibration:lights" + txt);
-
-      if (this.global.get_flowchart_data) {
-            return this.imagesEnabledPathToFileList(images);
-      }
-
-      var P = new ImageCalibration;
-      P.targetFrames = images; // [ enabled, path ];
-      P.enableCFA = this.is_color_files && this.local_debayer_pattern != 'None';
-      P.cfaPattern = this.global.debayerPattern_enums[this.global.debayerPattern_values.indexOf(this.local_debayer_pattern)];
-      if (this.par.auto_output_pedestal.val) {
-            console.writeln("runCalibrateLights, auto output pedestal");
-            P.outputPedestalMode = ImageCalibration.OutputPedestal_Auto;
-      } else {
-            console.writeln("runCalibrateLights, literal output pedestal " + this.par.output_pedestal.val);
-            P.outputPedestalMode = ImageCalibration.OutputPedestal_Literal;
-            P.outputPedestal = this.par.output_pedestal.val;
-      }
-      if (masterbiasPath != null) {
-            console.writeln("runCalibrateLights, master bias " + masterbiasPath);
-            P.masterBiasEnabled = true;
-            P.masterBiasPath = this.matchMasterToImages(images, masterbiasPath);
-      } else {
-            console.writeln("runCalibrateLights, no master bias");
-            P.masterBiasEnabled = false;
-            P.masterBiasPath = "";
-      }
-      if (masterdarkPath != null) {
-            console.writeln("runCalibrateLights, master dark " + masterdarkPath);
-            P.masterDarkEnabled = true;
-            P.masterDarkPath = this.matchMasterToImages(images, masterdarkPath);
-      } else {
-            console.writeln("runCalibrateLights, no master dark");
-            P.masterDarkEnabled = false;
-            P.masterDarkPath = "";
-      }
-      if (masterflatPath != null) {
-            console.writeln("runCalibrateLights, master flat " + masterflatPath);
-            P.masterFlatEnabled = true;
-            P.masterFlatPath = this.matchMasterToImages(images, masterflatPath);
-      } else {
-            console.writeln("runCalibrateLights, no master flat");
-            P.masterFlatEnabled = false;
-            P.masterFlatPath = "";
-      }
-      if (this.par.optimize_darks.val) {
-            P.calibrateBias = false;
-            if (this.darkIsBiasCalibrated(masterdarkPath)) {
-                  P.calibrateDark = false;
-            } else {
-                  P.calibrateDark = true;
-            }
-            P.calibrateFlat = false;
-            P.optimizeDarks = true;
-
-      } else {
-            P.calibrateBias = false;
-            P.calibrateDark = false;
-            P.calibrateFlat = false;
-            P.optimizeDarks = false;
-      }
-      P.outputDirectory = this.global.outputRootDir + this.global.AutoCalibratedDir;
-      P.overwriteExistingFiles = true;
-
-      P.executeGlobal();
-
-      this.printAndSaveProcessValues(P, "lights_" + filterName);
-      this.engine_end_process(node);
-
-      return this.fileNamesFromOutputData(P.outputData);
 }
 
 filesForImageIntegration(fileNames)
@@ -4138,27 +3630,6 @@ getFileKeywords(filePath)
       return keywords;
 }
 
-// Returns true if the master dark file had bias subtracted during its creation,
-// detected via the CALSTAT FITS keyword ('B' present). Falls back to the
-// pre_calibrate_darks parameter when the keyword is absent (e.g. older files).
-darkIsBiasCalibrated(darkPath)
-{
-      if (darkPath == null) {
-            return this.par.pre_calibrate_darks.val;
-      }
-      var keywords = this.getFileKeywords(darkPath);
-      for (var i = 0; i < keywords.length; i++) {
-            if (keywords[i].name == "CALSTAT") {
-                  var calstat = keywords[i].strippedValue.trim().toUpperCase();
-                  console.writeln("darkIsBiasCalibrated, CALSTAT=" + calstat + " in " + darkPath);
-                  return calstat.indexOf('B') >= 0;
-            }
-      }
-      // No CALSTAT keyword — fall back to user parameter.
-      console.writeln("darkIsBiasCalibrated, no CALSTAT in " + darkPath + ", using pre_calibrate_darks=" + this.par.pre_calibrate_darks.val);
-      return this.par.pre_calibrate_darks.val;
-}
-
 // Get filter keywpord for image. If filter is not found from FILTER keyword
 // or file name, default to color files.
 // Parameters:
@@ -5117,7 +4588,7 @@ runPixelMathSingleMappingEx(id, reason, mapping, createNewImage, symbols, rescal
       } else {
             var new_win = idWin;
       }
-      this.printAndSaveProcessValues(P, reason);
+      this.printAndSaveProcessValues(P, reason, idWin.mainView.id);
       this.engine_end_process(node, new_win, "PixelMath:" + reason, false);
 
       this.setAutoIntegrateVersionIfNeeded(this.util.findWindow(P.newImageId));
@@ -5170,7 +4641,7 @@ runPixelMathRGBMapping(newId, idWin, mapping_R, mapping_G, mapping_B)
       P.executeOn(idWin.mainView);
       idWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, "RGB");
+      this.printAndSaveProcessValues(P, "RGB", idWin.mainView.id);
       if (newId != null) {
             var new_win = this.util.findWindow(newId);
       } else {
@@ -5233,7 +4704,7 @@ runPixelMathRGBMappingFindRef(newId, mapping_R, mapping_G, mapping_B, channels_f
       P.executeOn(idWin.mainView);
       idWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, "RGB");
+      this.printAndSaveProcessValues(P, "RGB", idWin.mainView.id);
       var new_win = this.util.findWindow(newId);
       if (this.global.pixinsight_version_num >= 1080902) {
             new_win.copyAstrometricSolution(idWin);
@@ -5994,7 +5465,7 @@ removeStars(imgWin, linear_data, save_stars, save_array, stars_image_name, use_u
             var succp = P.executeOn(imgWin.mainView, false);
             
             imgWin.mainView.endProcess();
-            this.printAndSaveProcessValues(P);
+            this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
       } else {
             var succp = true;
       }
@@ -7151,7 +6622,7 @@ linearFitImage(refViewId, targetId, add_to_flowchart = false)
       P.executeOn(targetWin.mainView);
       targetWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(targetId));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(targetId), targetWin.mainView.id);
 
       this.util.setFITSKeyword(targetWin, "AutoLinearfit", "true", "Linear fit was done by AutoIntegrate");
       var refWin = ImageWindow.windowById(refViewId);
@@ -8188,7 +7659,7 @@ runABEex(win, replaceTarget, postfix, skip_flowchart, degree = null, normalize =
 
       win.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(win.mainView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(win.mainView.id), win.mainView.id);
       this.engine_end_process(null);
 
       this.util.addScriptWindow(GC_id);
@@ -8257,7 +7728,7 @@ runGCProcess(win, replaceTarget, postfix, from_lights)
             }
       }
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(win.mainView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(win.mainView.id), win.mainView.id);
 
       this.engine_end_process(node, this.util.findWindow(GC_id), "GradientCorrection");
 
@@ -8324,7 +7795,7 @@ runSpectrophotometricFluxCalibration(win)
             this.util.addCriticalStatus("SpectrophotometricFluxCalibration failed");
       }
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", win.mainView.id);
       this.engine_end_process(node, win, "SpectrophotometricFluxCalibration");
 }
 
@@ -8375,7 +7846,7 @@ runMultiscaleGradientCorrectionProcess(win)
                   this.util.copyWindowEx(win, win.mainView.id + "_MGC_after", true);
             }
       }
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(win.mainView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(win.mainView.id), win.mainView.id);
       this.engine_end_process(node, win, "MultiscaleGradientCorrection");
 
       return succp;
@@ -8676,7 +8147,7 @@ applySTF(imgView, stf, iscolor, save_process = false)
       imgView.endProcess();
 
       if (save_process) {
-            this.printAndSaveProcessValues(HT, "autostf_" + this.findChannelFromNameIf(imgView.id));
+            this.printAndSaveProcessValues(HT, "autostf_" + this.findChannelFromNameIf(imgView.id), imgView.id);
       }     
       this.engine_end_process(null);
 }
@@ -8826,7 +8297,7 @@ runHistogramTransformMultiscaleAdaptiveStretch(GC_win, image_stretching)
       P.executeOn(GC_win.mainView);
       GC_win.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, "MultiscaleAdaptiveStretch");
+      this.printAndSaveProcessValues(P, "MultiscaleAdaptiveStretch", GC_win.mainView.id);
       this.engine_end_process(null);
 
       return GC_win;
@@ -8857,7 +8328,7 @@ runHistogramTransformMaskedStretch(GC_win, image_stretching, histogram_prestretc
 
       GC_win.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(GC_win.mainView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(GC_win.mainView.id), GC_win.mainView.id);
       this.engine_end_process(null);
 
       return GC_win;
@@ -8911,7 +8382,7 @@ runHistogramTransformArcsinhStretch(GC_win, image_stretching, histogram_poststre
 
             GC_win.mainView.endProcess();
 
-            this.printAndSaveProcessValues(P, this.findChannelFromNameIf(GC_win.mainView.id));
+            this.printAndSaveProcessValues(P, this.findChannelFromNameIf(GC_win.mainView.id), GC_win.mainView.id);
             this.engine_end_process(null);
             this.util.runGarbageCollection();
 
@@ -9310,7 +8781,7 @@ stretchHistogramTransformAdjustShadows(new_win, channel, use_median)
       P.executeOn(new_win.mainView);
       new_win.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, "shadows_" + this.findChannelFromNameIf(new_win.mainView.id));
+      this.printAndSaveProcessValues(P, "shadows_" + this.findChannelFromNameIf(new_win.mainView.id), new_win.mainView.id);
 
       this.engine_end_process(null);
 
@@ -9749,31 +9220,31 @@ runHistogramTransform(GC_win, iscolor, type)
             GC_win = this.runHistogramTransformMaskedStretch(GC_win, image_stretching, true);
 
       } else if (image_stretching == 'Arcsinh Stretch') {
-            this.skip_process_value_save = true;
+            this.global.skip_process_value_save = true;
             GC_win = this.runHistogramTransformArcsinhStretch(GC_win, image_stretching, false);
 
       } else if (image_stretching == 'Asinh+Histogram stretch') {
-            this.skip_process_value_save = true;
+            this.global.skip_process_value_save = true;
             GC_win = this.runHistogramTransformArcsinhStretch(GC_win, image_stretching, true);
 
       } else if (image_stretching == 'Histogram stretch') {
-            this.skip_process_value_save = true;
+            this.global.skip_process_value_save = true;
             GC_win = this.stretchHistogramTransformIterations(GC_win, iscolor, image_stretching, this.par.histogram_stretch_target.val, null);
 
       } else if (image_stretching == 'Logarithmic stretch') {
-            this.skip_process_value_save = true;
+            this.global.skip_process_value_save = true;
             GC_win = this.stretchFunctionIterations(GC_win, iscolor, image_stretching, this.par.other_stretch_target.val, 10);
 
       } else if (image_stretching == 'Square root stretch') {
-            this.skip_process_value_save = true;
+            this.global.skip_process_value_save = true;
             GC_win = this.stretchFunctionIterations(GC_win, iscolor, image_stretching, this.par.other_stretch_target.val, 1);
 
       } else if (image_stretching == 'Shadow stretch') {
-            this.skip_process_value_save = true;
+            this.global.skip_process_value_save = true;
             GC_win = this.stretchFunctionIterations(GC_win, iscolor, image_stretching, this.par.other_stretch_target.val, 1);
 
       } else if (image_stretching == 'Highlight stretch') {
-            this.skip_process_value_save = true;
+            this.global.skip_process_value_save = true;
             GC_win = this.stretchFunctionIterations(GC_win, iscolor, image_stretching, this.par.other_stretch_target.val, 1);
 
       } else if (image_stretching == 'Histogram direct') {
@@ -9782,7 +9253,7 @@ runHistogramTransform(GC_win, iscolor, type)
       } else {
             this.util.throwFatalError("Bad image stretching value " + image_stretching + " with type " + type);
       }
-      this.skip_process_value_save = false;
+      this.global.skip_process_value_save = false;
       this.util.setFITSKeyword(GC_win, "AutoIntegrateNonLinear", image_stretching, "");
       this.engine_end_process(node, GC_win, image_stretching);
 
@@ -9832,7 +9303,7 @@ runACDNRReduceNoise(imgWin, maskWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
 
       this.engine_end_process(node, imgWin, "ACDNR:noise");
 
@@ -9976,7 +9447,7 @@ runMultiscaleLinearTransformReduceNoise(imgWin, maskWin, strength)
             imgWin.removeMask();
       }
 
-      this.printAndSaveProcessValues(P, "noise_" + this.findChannelFromNameIf(imgWin.mainView.id));
+      this.printAndSaveProcessValues(P, "noise_" + this.findChannelFromNameIf(imgWin.mainView.id), imgWin.mainView.id);
       this.engine_end_process(node, imgWin, "MultiscaleLinearTransform:noise");
 
       imgWin.mainView.endProcess();
@@ -10065,7 +9536,7 @@ runBlurXTerminator(imgWin, correct_only, for_image_solver = false)
       
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id), imgWin.mainView.id);
       this.engine_end_process(node, imgWin, "BlurXTerminator");
 }
 
@@ -10118,7 +9589,7 @@ runNoiseXTerminator(imgWin, linear)
       
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id), imgWin.mainView.id);
       this.engine_end_process(node, imgWin, "NoiseXTerminator");
 }
 
@@ -10155,7 +9626,7 @@ runDeepSNR(imgWin, linear)
       
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id), imgWin.mainView.id);
       this.engine_end_process(node, imgWin, "DeepSNR");
 }
 
@@ -10232,7 +9703,7 @@ runColorReduceNoise(imgWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id), imgWin.mainView.id);
       this.engine_end_process(node, imgWin, "TGVDenoise");
 
       this.guiUpdatePreviewWin(imgWin);
@@ -10277,7 +9748,7 @@ starReduceNoise(imgWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
       this.engine_end_process(node, imgWin, "TGVDenoise");
 }
 
@@ -10570,7 +10041,7 @@ runBackgroundNeutralization(imgWin, find_true_background = true, skip_flowchart 
 
       imgView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgView.id);
       this.engine_end_process(node, imgWin, "BackgroundNeutralization");
 
       this.guiUpdatePreviewId(imgView.id);
@@ -10682,7 +10153,7 @@ runDBEprocess(imgWin, image_samples)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(imgWin.mainView.id), imgWin.mainView.id);
 
       this.engine_end_process(null);
 
@@ -11583,7 +11054,7 @@ runColorCalibrationProcess(imgWin, roi)
 
             imgWin.mainView.endProcess();
 
-            this.printAndSaveProcessValues(P);
+            this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
             this.engine_end_process(node, imgWin, "ColorCalibration");
 
             this.guiUpdatePreviewId(imgWin.mainView.id);
@@ -11700,7 +11171,7 @@ runSPCC(imgWin, phase)
                   this.util.throwFatalError("SpectrophotometricColorCalibration failed");
             }
 
-            this.printAndSaveProcessValues(P);
+            this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
             this.engine_end_process(node, imgWin, "SpectrophotometricColorCalibration");
 
             this.guiUpdatePreviewId(imgWin.mainView.id);
@@ -11794,7 +11265,7 @@ runColorSaturation(imgWin, maskWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, "colorsaturation");
+      this.printAndSaveProcessValues(P, "colorsaturation", imgWin.mainView.id);
       this.engine_end_process(node);
 
       this.guiUpdatePreviewWin(imgWin);
@@ -11847,7 +11318,7 @@ runCurvesTransformationSaturation(imgWin, maskWin, increase = true)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, "saturation");
+      this.printAndSaveProcessValues(P, "saturation", imgWin.mainView.id);
       this.engine_end_process(node, imgWin, "CurvesTransformation:saturation");
 
       this.guiUpdatePreviewWin(imgWin);
@@ -11898,7 +11369,7 @@ runCurvesTransformationChrominance(imgWin, maskWin, increase = true)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, "chrominance");
+      this.printAndSaveProcessValues(P, "chrominance", imgWin.mainView.id);
       this.engine_end_process(node, imgWin, "CurvesTransformation:chrominance");
 
       this.guiUpdatePreviewWin(imgWin);
@@ -11974,7 +11445,7 @@ runLRGBCombination(RGB_id, L_id)
 
       RGBimgView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", RGBimgView.id);
       this.engine_end_process(node, targetWin, "LRGBCombination");
 
       this.guiUpdatePreviewId(RGBimgView.id);
@@ -12009,7 +11480,7 @@ runSCNR(imgWin, fixing_stars)
 
       RGBimgView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", RGBimgView.id);
       this.engine_end_process(node, imgWin, "SCNR");
 
       this.guiUpdatePreviewId(RGBimgView.id);
@@ -12082,7 +11553,7 @@ runMultiscaleLinearTransformSharpen(imgWin, maskWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P, "sharpen");
+      this.printAndSaveProcessValues(P, "sharpen", imgWin.mainView.id);
       this.engine_end_process(node, imgWin, "MultiscaleLinearTransform:sharpen");
 
       this.guiUpdatePreviewWin(imgWin);
@@ -12760,11 +12231,6 @@ getFileProcessedStatus(fileNames, postfix)
       return this.getFileProcessedStatusEx(fileNames, postfix, this.global.outputRootDir + this.global.AutoOutputDir);
 }
 
-getFileProcessedStatusCalibrated(fileNames, postfix)
-{
-      return this.getFileProcessedStatusEx(fileNames, postfix, this.global.outputRootDir + this.global.AutoCalibratedDir);
-}
-
 preprocessedName(pi)
 {
       switch (pi) {
@@ -13064,11 +12530,11 @@ createChannelImages(parent, auto_continue)
 
             if (this.par.generate_masters_only.val) {
                   // Generate master calibration files only, no lights needed
-                  this.util.setOutputRootDir(this.getOutputDirFromCalibrationFiles());
+                  this.util.setOutputRootDir(this.calibrate.getOutputDirFromCalibrationFiles());
                   this.util.ensureDir(this.global.outputRootDir);
                   this.util.ensureDir(this.util.combinePath(this.global.outputRootDir, this.global.AutoOutputDir));       // For temp files during processing
                   this.util.ensureDir(this.util.combinePath(this.global.outputRootDir, this.global.AutoProcessedDir));    // For AutoIntegrate.log
-                  this.calibrateEngine(null);
+                  this.calibrate.calibrateEngine(null);
                   return this.retval.INCOMPLETE;
             }
 
@@ -13172,7 +12638,7 @@ createChannelImages(parent, auto_continue)
              *
              * Run image calibration if we have calibration frames.
              *********************************************************************/
-            var calibrateInfo = this.calibrateEngine(filtered_lights);
+            var calibrateInfo = this.calibrate.calibrateEngine(filtered_lights);
             this.lightFileNames = calibrateInfo[0];
             filename_postfix = filename_postfix + calibrateInfo[1];
             this.guiUpdatePreviewFilename(this.lightFileNames[0]);
@@ -13597,7 +13063,7 @@ createNewTempMaskFromLinearWin(imgWin, is_color)
 {
       console.writeln("createNewTempMaskFromLinearWin from " + imgWin.mainView.id);
       this.flowchart.flowchartMaskBegin("New mask");
-      this.creating_mask = true;
+      this.global.creating_mask = true;
 
       var winCopy = this.util.copyWindowEx(imgWin, imgWin.mainView.id + "_tmp", true);
 
@@ -13610,7 +13076,7 @@ createNewTempMaskFromLinearWin(imgWin, is_color)
 
       this.util.closeOneWindow(winCopy);
       this.flowchart.flowchartMaskEnd("New mask");
-      this.creating_mask = false;
+      this.global.creating_mask = false;
 
       return maskWin;
 }
@@ -13637,7 +13103,7 @@ LRGBEnsureMaskEx(L_id_for_mask, stretched)
       } else {
             var L_win;
             this.flowchart.flowchartMaskBegin("New mask");
-            this.creating_mask = true;
+            this.global.creating_mask = true;
             if (L_id_for_mask != null) {
                   this.util.addProcessingStep("Using image " + L_id_for_mask + " for a mask");
                   L_win = this.util.copyWindowEx(ImageWindow.windowById(L_id_for_mask), this.ppar.win_prefix + "L_win_mask", true);
@@ -13673,7 +13139,7 @@ LRGBEnsureMaskEx(L_id_for_mask, stretched)
             this.mask_win = this.newMaskWindow(L_win, this.mask_win_id, false);
             this.util.closeOneWindowById(L_win.mainView.id);
             this.flowchart.flowchartMaskEnd("New mask");
-            this.creating_mask = false;
+            this.global.creating_mask = false;
       }
 }
 
@@ -13700,7 +13166,7 @@ colorEnsureMask(color_img_id, RGBstretched, force_new_mask)
             this.mask_win = this.range_mask_win;
       } else {
             this.flowchart.flowchartMaskBegin("New mask");
-            this.creating_mask = true;
+            this.global.creating_mask = true;
             var color_win = ImageWindow.windowById(color_img_id);
             this.util.addProcessingStep("Using image " + color_img_id + " for a mask");
             var color_win_copy = this.util.copyWindowEx(color_win, "color_win_mask", true);
@@ -13716,7 +13182,7 @@ colorEnsureMask(color_img_id, RGBstretched, force_new_mask)
             this.mask_win = this.newMaskWindow(color_win_copy, this.mask_win_id, false);
             this.util.closeOneWindowById(color_win_copy.mainView.id);
             this.flowchart.flowchartMaskEnd("New mask");
-            this.creating_mask = false;
+            this.global.creating_mask = false;
       }
       console.writeln("colorEnsureMask done");
 }
@@ -14064,7 +13530,7 @@ combineRGBimageEx(target_name, images)
       P.executeOn(win.mainView);
       win.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", win.mainView.id);
       this.engine_end_process(node, win, "ChannelCombination");
 
       if (this.global.pixinsight_version_num >= 1080902) {
@@ -14547,7 +14013,7 @@ create_continuum_subtracted_image(hrr_win)
       P.newImageSampleFormat = PixelMath.SameAsTarget;
 
       P.executeOn(hrr_win.mainView);
-      this.printAndSaveProcessValues(P, "continuum_subtracted");
+      this.printAndSaveProcessValues(P, "continuum_subtracted", hrr_win.mainView.id);
 }
 
 normalize_image(imgWin)
@@ -14581,7 +14047,7 @@ normalize_image(imgWin)
       P.newImageSampleFormat = PixelMath.SameAsTarget;
 
       P.executeOn(imgWin.mainView);
-      this.printAndSaveProcessValues(P, "normalize");
+      this.printAndSaveProcessValues(P, "normalize", imgWin.mainView.id);
 }
 
 convert_to_grayscale(imgWin)
@@ -14589,7 +14055,7 @@ convert_to_grayscale(imgWin)
       var P = new ConvertToGrayscale;
 
       P.executeOn(imgWin.mainView);
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
 }
 
 boost_Ha(imgWin, boost_factor)
@@ -14601,7 +14067,7 @@ boost_Ha(imgWin, boost_factor)
       P.useLightnessMask = true;
 
       P.executeOn(imgWin.mainView);
-      this.printAndSaveProcessValues(P, "boostHA");
+      this.printAndSaveProcessValues(P, "boostHA", imgWin.mainView.id);
 }
 
 HRR_stretch(imgWin, targetBackground)
@@ -15477,7 +14943,7 @@ invertImage(targetView)
       P.executeOn(targetView, true);
       targetView.endProcess();
 
-      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(targetView.id));
+      this.printAndSaveProcessValues(P, this.findChannelFromNameIf(targetView.id), targetView.id);
       this.engine_end_process(node);
 }
 
@@ -15510,7 +14976,7 @@ createStarFixMask(imgView)
       P.executeOn(imgView, false);
       imgView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgView.mainView.id);
       this.engine_end_process(node);
 
       this.star_fix_mask_win = ImageWindow.activeWindow;
@@ -15609,7 +15075,7 @@ enhancementsFixStarCores(targetWin)
 
       P.executeOn(targetWin.mainView, false);
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", targetWin.mainView.id);
 
       console.writeln("enhancementsFixStarCores:set star mask");
 
@@ -15636,7 +15102,7 @@ enhancementsFixStarCores(targetWin)
 
       P.executeOn(targetWin.mainView, false);
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", targetWin.mainView.id);
            
       // Cleanup
       targetWin.removeMask();
@@ -15724,7 +15190,7 @@ enhancementsDarkerBackground(imgWin, maskWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
       this.engine_end_process(node);
 
       // this.guiUpdatePreviewWin(imgWin);
@@ -15755,7 +15221,7 @@ enhancementsDarkerHighlights(imgWin, maskWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
       this.engine_end_process(node);
 
       // this.guiUpdatePreviewWin(imgWin);
@@ -15790,7 +15256,7 @@ enhancementsAdjustChannels(imgWin)
       P.executeOn(imgWin.mainView);
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
       this.engine_end_process(node);
 }
 
@@ -15840,7 +15306,7 @@ enhancementsHDRMultiscaleTransform(imgWin, maskWin)
 
             imgWin.mainView.endProcess();
 
-            this.printAndSaveProcessValues(P);
+            this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
             this.engine_end_process(node);
 
             if (this.par.enhancements_HDRMLT_color.val == 'Color corrected'
@@ -15928,7 +15394,7 @@ enhancementsLocalHistogramEqualization(imgWin, maskWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
       this.engine_end_process(node);
 
       // this.guiUpdatePreviewWin(imgWin);
@@ -15984,7 +15450,7 @@ enhancementsExponentialTransformation(imgWin, maskWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
       this.engine_end_process(node);
 
       // this.guiUpdatePreviewWin(imgWin);
@@ -16004,7 +15470,7 @@ createNewStarMaskWin(imgWin)
       P.executeOn(imgWin.mainView, false);
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
       this.engine_end_process(node);
 
       return ImageWindow.activeWindow;
@@ -16175,7 +15641,7 @@ enhancementsSmallerStars(imgWin, is_star_image)
 
       targetWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", targetWin.mainView.id);
       this.engine_end_process(node);
 
       // this.guiUpdatePreviewWin(targetWin);
@@ -16200,7 +15666,7 @@ enhancementsContrast(imgWin)
 
       imgWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", imgWin.mainView.id);
       this.engine_end_process(node);
 
       // this.guiUpdatePreviewWin(imgWin);
@@ -16380,7 +15846,7 @@ combineLowPassandHighPass(enhancementsWin, low_pass_win, high_pass_win)
 
       P.executeOn(enhancementsWin.mainView, false);
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", enhancementsWin.mainView.id);
       this.engine_end_process(node);
 }
 
@@ -16441,7 +15907,7 @@ enhancementsHighPassSharpen(enhancementsWin, mask_win)
 
       P.executeOn(low_pass_win.mainView, false);
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", low_pass_win.mainView.id);
       this.engine_end_process(node);
 
       // ---------------------------------------------
@@ -16462,7 +15928,7 @@ enhancementsHighPassSharpen(enhancementsWin, mask_win)
 
       P.executeOn(high_pass_win.mainView, false);
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", high_pass_win.mainView.id);
       this.engine_end_process(node);
 
       if (this.par.enhancements_highpass_sharpen_noise_reduction.val) {
@@ -16522,7 +15988,7 @@ enhancementsUnsharpMask(enhancementsWin, mask_win)
       P.executeOn(enhancementsWin.mainView, false);
       enhancementsWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", enhancementsWin.mainView.id);
       this.engine_end_process(node);
 
       if (this.mask_win != null) {
@@ -16556,7 +16022,7 @@ enhancementsClarity(enhancementsWin, mask_win)
       P.executeOn(enhancementsWin.mainView, false);
       enhancementsWin.mainView.endProcess();
 
-      this.printAndSaveProcessValues(P);
+      this.printAndSaveProcessValues(P, "", enhancementsWin.mainView.id);
       this.engine_end_process(node);
 
       if (mask_win != null) {
@@ -17361,10 +16827,10 @@ enhancementsProcessing(parent, id, apply_directly)
                   this.mask_win_id = this.ppar.win_prefix + "AutoMask";
                   this.util.closeOneWindowById(this.mask_win_id);
                   this.flowchart.flowchartMaskBegin("New mask:enhancements");
-                  this.creating_mask = true;
+                  this.global.creating_mask = true;
                   this.mask_win = this.newMaskWindow(enhancementsWin, this.mask_win_id, false);
                   this.flowchart.flowchartMaskEnd("New mask:enhancements");
-                  this.creating_mask = false;
+                  this.global.creating_mask = false;
             }
             console.writeln("Use mask " + this.mask_win.mainView.id);
       }
@@ -18108,7 +17574,7 @@ cropImageIf(id, show_in_flowchart = true)
       // Add keyword to indicate that the image has been cropped
       this.util.setFITSKeyword(window, "AutoCrop", "true", "Image cropped by AutoIntegrate");
 
-      this.printAndSaveProcessValues(crop, this.findChannelFromNameIf(window.mainView.id));
+      this.printAndSaveProcessValues(crop, this.findChannelFromNameIf(window.mainView.id), window.mainView.id);
       this.engine_end_process(node, window, "Crop");
 
       if (this.par.save_cropped_images.val) {
@@ -18352,424 +17818,6 @@ createCropInformationAutoContinue()
       }
       this.findCropInformationAutoContinue();
 }
-
-getOutputDirFromCalibrationFiles()
-{
-      // Use the first available calibration file to determine output directory
-      var calibFile = null;
-      if (this.biasFileNames && this.biasFileNames.length > 0) {
-            calibFile = this.biasFileNames[0];
-      } else if (this.darkFileNames && this.darkFileNames.length > 0) {
-            calibFile = this.darkFileNames[0];
-      } else if (this.flatFileNames && this.flatFileNames.length > 0) {
-            calibFile = this.flatFileNames[0];
-      } else if (this.flatdarkFileNames && this.flatdarkFileNames.length > 0) {
-            calibFile = this.flatdarkFileNames[0];
-      }
-      if (calibFile != null) {
-            return this.util.getOutputDir(calibFile);
-      }
-      return this.global.outputRootDir;
-}
-
-/***************************************************************************
- *
- *    this.calibrateEngine
- *
- * Calibration this to run image calibration
- * if bias, dark and/or flat files are selected.
- */
-calibrateEngine(filtered_lights)
-{
-       if (this.biasFileNames == null) {
-             this.biasFileNames = [];
-       }
-       if (this.flatdarkFileNames == null) {
-             this.flatdarkFileNames = [];
-       }
-       if (this.flatFileNames == null) {
-             this.flatFileNames = [];
-       }
-       if (this.darkFileNames == null) {
-             this.darkFileNames = [];
-       }
-       if (this.lightFileNames == null) {
-             this.lightFileNames = [];
-       }
-       if (this.biasFileNames.length == 0
-           && this.flatdarkFileNames.length == 0
-           && this.flatFileNames.length == 0
-           && this.darkFileNames.length == 0)
-       {
-             // do not calibrate
-             this.util.addProcessingStep("calibrateEngine, no bias, flat or dark files");
-             return [ this.lightFileNames , '' ];
-       }
- 
-       this.util.addProcessingStepAndStatusInfo("Image calibration");
- 
-       this.util.ensureDir(this.global.outputRootDir);
-       this.util.ensureDir(this.util.combinePath(this.global.outputRootDir, this.global.AutoMasterDir));
-       this.util.ensureDir(this.util.combinePath(this.global.outputRootDir, this.global.AutoOutputDir));
-       if (!this.par.generate_masters_only.val) {
-            // When we generate master we still use
-            // - AutoOutputDir for temp files
-            // - AutoProcessedDir for AutoIntegrate.log
-            this.util.ensureDir(this.util.combinePath(this.global.outputRootDir, this.global.AutoCalibratedDir));
-       }
- 
-       if (this.global.get_flowchart_data) {
-            // Filter files for this.global.get_flowchart_data.
-            this.flatFileNames = this.flowchartFilterFiles(this.flatFileNames, this.global.pages.FLATS);
-            this.flatdarkFileNames = this.flowchartFilterFiles(this.flatdarkFileNames, this.global.pages.FLAT_DARKS);
-            // Dark files are not filtered here: the dark integration code calls
-            // groupDarksByExposureTime internally, so each exposure group already
-            // produces its own this.flowchart operation without further filtering needed.
-      }
-
-       // Collect filter files
-       var filtered_flats = this.getFilterFiles(this.flatFileNames, this.global.pages.FLATS, '');
-       var filtered_flatdarks = this.getFilterFiles(this.flatdarkFileNames, this.global.pages.FLAT_DARKS, '');
- 
-       this.is_color_files = filtered_flats.color_files;
- 
-       if (this.flatFileNames.length > 0 && this.lightFileNames.length > 0) {
-             // we have flats and lights
-             // check that filtered files match
-             for (var i = 0; i < filtered_flats.allfilesarr.length; i++) {
-                   var is_flats = filtered_flats.allfilesarr[i].files.length > 0;
-                   var is_lights = filtered_lights.allfilesarr[i].files.length > 0;
-                   if (is_lights && !is_flats) {
-                         // We need to have flafs for all filters used in lights
-                         this.util.throwFatalError("No flats for filter " + filtered_flats.allfilesarr[i].filter + " but we have lights for this filter. Please check that you have flats for all filters used in lights and that the filter information is correct.");
-                   }
-             }
-       }
-       if (this.flatFileNames.length > 0 && this.flatdarkFileNames.length > 0) {
-            // we have flats and flat darks
-            // check that filtered files match
-            for (var i = 0; i < filtered_flats.allfilesarr.length; i++) {
-                  var is_flats = filtered_flats.allfilesarr[i].files.length > 0;
-                  var is_flatdarks = filtered_flatdarks.allfilesarr[i].files.length > 0;
-                  if (is_flats != is_flatdarks) {
-                        // lights and flats do not match on filters
-                        this.util.throwFatalError("Filters on flats and flat darks do not match for filter " + filtered_flats.allfilesarr[i].filter + ". Please check that you have flat darks for all filters used in flats and that the filter information is correct.");
-                  }
-            }
-      }
-
-       /*
-        * Bias files
-        *
-        * Generate a master bias file.
-        */
-       if (this.par.bias_master_files.val) {
-             // We have an array of master bias files, we match them by resolution
-             this.util.addProcessingStep("calibrateEngine use existing master bias files " + this.biasFileNames);
-             var masterbiasPath = this.biasFileNames;
-       } else if (this.biasFileNames.length == 1) {
-             this.util.addProcessingStep("calibrateEngine use existing master bias " + this.biasFileNames[0]);
-             var masterbiasPath = this.biasFileNames[0];
-       } else if (this.biasFileNames.length > 0) {
-             this.util.addProcessingStep("calibrateEngine generate master bias using " + this.biasFileNames.length + " files");
-             // integrate bias images
-             var biasimages = this.filesForImageIntegration(this.biasFileNames);
-             var masterbiasid = this.runImageIntegrationBiasDarks(biasimages, this.ppar.win_prefix + "AutoMasterBias", "bias");
- 
-             // save master bias
-             this.setImagetypKeyword(this.util.findWindow(masterbiasid), "Master bias");
-             var masterbiasPath = this.saveMasterWindow(this.global.outputRootDir, masterbiasid);
- 
-             // optionally generate superbias
-             if (this.par.create_superbias.val) {
-                   var masterbiaswin = this.util.findWindow(masterbiasid);
-                   var mastersuperbiasid = this.runSuberBias(masterbiaswin);
-                   this.setImagetypKeyword(this.util.findWindow(mastersuperbiasid), "Master bias");
-                   var masterbiasPath = this.saveMasterWindow(this.global.outputRootDir, mastersuperbiasid);
-                   this.guiUpdatePreviewId(mastersuperbiasid);
-             } else {
-                   this.guiUpdatePreviewId(masterbiasid);
-             }
-       } else {
-             this.util.addProcessingStep("calibrateEngine no master bias");
-             var masterbiasPath = null;
-       }
- 
-       /*
-        * Flat dark files
-        *
-        * Generate a master flat dark file for each filter.
-        */
-       if (this.flatdarkFileNames.length > 0) {
-            // generate master flat dark for each filter
-            this.util.addProcessingStep("calibrateEngine generate master flat darks");
-            var masterflatdarkPath = [];
-            this.flowchart.flowchartParentBegin("Flat darks");
-            for (var i = 0; i < filtered_flatdarks.allfilesarr.length; i++) {
-                  var filterFiles = filtered_flatdarks.allfilesarr[i].files;
-                  var filterName = filtered_flatdarks.allfilesarr[i].filter;
-                  this.flowchart.flowchartChildBegin(filterName);
-                  if (this.par.flat_dark_master_files.val) {
-                        // We have an array of master flat dark files, we match them by resolution
-                        this.util.addProcessingStep("calibrateEngine use existing " + filterName + " master flat dark files");
-                        masterflatdarkPath[i] = filterFiles.length == 1 ? filterFiles[0].name : filterFiles.map(function(f) { return f.name; });
-                  } else if (filterFiles.length == 1) {
-                        this.util.addProcessingStep("calibrateEngine use existing " + filterName + " master flat dark " + filterFiles[0].name);
-                        masterflatdarkPath[i] = filterFiles[0].name;
-                  } else if (filterFiles.length > 0) {
-                        // integrate flat darks to generate master flat dark for each filter
-                        this.util.addProcessingStep("calibrateEngine create " + filterName + " master flat dark");
-                        var flatdarkimages = this.filesForImageIntegrationFromFilearr(filterFiles);
-                        console.writeln("flatdarkimages[0] " + flatdarkimages[0][1]);
-                        var masterflatdarkid = this.runImageIntegrationBiasDarks(flatdarkimages, this.ppar.win_prefix + "AutoMasterFlatDark_" + filterName, "flatdark");
-                        console.writeln("masterflatdarkid " + masterflatdarkid);
-                        this.setImagetypKeyword(this.util.findWindow(masterflatdarkid), "Master flat dark");
-                        masterflatdarkPath[i] = this.saveMasterWindow(this.global.outputRootDir, masterflatdarkid);
-                        this.guiUpdatePreviewId(masterflatdarkid);
-                  } else {
-                        masterflatdarkPath[i] = null;
-                  }
-                  this.flowchart.flowchartChildEnd(filterName);
-            }
-            this.flowchart.flowchartParentEnd("Flat darks");
-      } else {
-             this.util.addProcessingStep("calibrateEngine no master flat dark");
-             var masterflatdarkPath = null;
-       }
- 
-       /*
-        * Dark files
-        *
-        * Generate master dark file(s). If we have dark files with different exposure times, 
-        * we generate a master dark for each exposure time group. If we have only one dark file, 
-        * we use it as master dark file assuming it is already calibrated. 
-        * If we have master dark files option selected, we use the provided master dark files.
-        * If we have only one dark file for each exposure, we use it as master dark file assuming 
-        * it is already calibrated. 
-        * If we have multiple dark files, we group them by exposure time and generate a master dark 
-        * for each exposure time group.
-        */
-       if (this.par.dark_master_files.val) {
-             this.util.addProcessingStep("calibrateEngine use existing master dark files " + this.darkFileNames);
-             if (this.darkFileNames.length == 1) {
-                   var masterdarkPath = this.darkFileNames[0];
-             } else {
-                   // Group master darks by exposure time so selectMasterDarkForExptime can match them
-                   var darkExptimeGroups = this.groupDarksByExposureTime(this.darkFileNames);
-                   if (darkExptimeGroups.length == 1) {
-                         // All same exposure time, pass as array for resolution matching
-                         var masterdarkPath = this.darkFileNames;
-                   } else {
-                         // Multiple exposure time groups, build { exptime, path } array
-                         var masterdarkPath = [];
-                         for (var dg = 0; dg < darkExptimeGroups.length; dg++) {
-                               var groupFiles = darkExptimeGroups[dg].files;
-                               var groupExptime = darkExptimeGroups[dg].exptime;
-                               if (groupFiles.length == 1) {
-                                     masterdarkPath.push({ exptime: groupExptime, path: groupFiles[0] });
-                               } else {
-                                     // Multiple master darks at same exposure, pass array for resolution matching
-                                     masterdarkPath.push({ exptime: groupExptime, path: groupFiles });
-                               }
-                         }
-                   }
-             }
-       } else if (this.darkFileNames.length == 1) {
-             this.util.addProcessingStep("calibrateEngine use existing master dark " + this.darkFileNames[0]);
-             var masterdarkPath = this.darkFileNames[0];
-       } else if (this.darkFileNames.length > 0) {
-             // Group dark frames by exposure time
-             var darkExptimeGroups = this.groupDarksByExposureTime(this.darkFileNames);
-             this.util.addProcessingStep("calibrateEngine generate master dark using " + this.darkFileNames.length + " files in " + darkExptimeGroups.length + " exposure time group(s)");
-
-             if (darkExptimeGroups.length == 1) {
-                   // Single exposure time group, create one master dark as before
-                   var groupExptime = darkExptimeGroups[0].exptime;
-                   var groupName = groupExptime + "s";
-                   if (this.par.pre_calibrate_darks.val && masterbiasPath != null) {
-                         var darkcalFileNames = this.runCalibrateDarks(darkExptimeGroups[0].files, masterbiasPath);
-                         var darkimages = this.filesForImageIntegration(darkcalFileNames);
-                   } else {
-                         var darkimages = this.filesForImageIntegration(darkExptimeGroups[0].files);
-                   }
-                   var masterdarkid = this.runImageIntegrationBiasDarks(darkimages, this.ppar.win_prefix + "AutoMasterDark_" + groupName, "dark", groupExptime);
-                   let win = this.util.findWindow(masterdarkid);
-                   this.util.setFITSKeyword(win, "EXPTIME", groupExptime.toString(), "Exposure time for master dark");
-                   this.setImagetypKeyword(win, "Master dark");
-                   if (this.par.pre_calibrate_darks.val && masterbiasPath != null) {
-                         this.util.setFITSKeyword(win, "CALSTAT", "B", "Calibration status: B=bias subtracted");
-                   }
-                   var masterdarkPath = this.saveMasterWindow(this.global.outputRootDir, masterdarkid);
-                   this.guiUpdatePreviewId(masterdarkid);
-             } else {
-                   // Multiple exposure time groups, create a master dark for each group
-                   var masterdarkPath = []; // array of { exptime, path }
-                   this.flowchart.flowchartParentBegin("Darks");
-                   for (var dg = 0; dg < darkExptimeGroups.length; dg++) {
-                         var groupExptime = darkExptimeGroups[dg].exptime;
-                         var groupFiles = darkExptimeGroups[dg].files;
-                         var groupName = groupExptime + "s";
-                         this.flowchart.flowchartChildBegin(groupName);
-
-                         if (groupFiles.length == 1) {
-                               // Single file in group, treat as pre-made master dark
-                               this.util.addProcessingStep("calibrateEngine use existing master dark for " + groupName + ": " + groupFiles[0]);
-                               masterdarkPath.push({ exptime: groupExptime, path: groupFiles[0] });
-                         } else {
-                               this.util.addProcessingStep("calibrateEngine create master dark for " + groupName + " using " + groupFiles.length + " files");
-
-                               if (this.par.pre_calibrate_darks.val && masterbiasPath != null) {
-                                     var darkcalFileNames = this.runCalibrateDarks(groupFiles, masterbiasPath);
-                                     var darkimages = this.filesForImageIntegration(darkcalFileNames);
-                               } else {
-                                     var darkimages = this.filesForImageIntegration(groupFiles);
-                               }
-                               var masterdarkid = this.runImageIntegrationBiasDarks(darkimages, this.ppar.win_prefix + "AutoMasterDark_" + groupName, "dark", groupExptime);
-                               let win = this.util.findWindow(masterdarkid);
-                               this.setImagetypKeyword(win, "Master dark");
-                               this.util.setFITSKeyword(win, "EXPTIME", groupExptime.toString(), "Exposure time for master dark");
-                               if (this.par.pre_calibrate_darks.val && masterbiasPath != null) {
-                                     this.util.setFITSKeyword(win, "CALSTAT", "B", "Calibration status: B=bias subtracted");
-                               }
-                               var groupMasterPath = this.saveMasterWindow(this.global.outputRootDir, masterdarkid);
-                               this.guiUpdatePreviewId(masterdarkid);
-                               masterdarkPath.push({ exptime: groupExptime, path: groupMasterPath });
-                         }
-
-                         this.flowchart.flowchartChildEnd(groupName);
-                   }
-                   this.flowchart.flowchartParentEnd("Darks");
-             }
-       } else {
-             this.util.addProcessingStep("calibrateEngine no master dark");
-             var masterdarkPath = null;
-       }
- 
-       /* 
-        * Flat files
-        *    
-        * Generate master flat files for each filter. If we have master flat files option selected, 
-        * we use the provided master flat files.
-        * If we have only one flat file for each filter, we use it as master flat file assuming 
-        * it is already calibrated.
-        * If we have multiple flat files for a filter, we calibrate them with master bias and/or 
-        * master dark and/or master flat dark if available and then integrate to generate a master 
-        * flat for the filter.
-        * If we have flat darks, we use them for flat calibration. 
-        * If we do not have flat darks, we calibrate flats using bias.
-        * Optionally, if we have darks but no flat darks, we use master darks for flat calibration. We select the 
-        * master dark with the closest exposure time to the flat frames.
-        */
-       this.util.addProcessingStepAndStatusInfo("Image calibration, generate master flats");
-       var masterflatPath = [];
-       this.flowchart.flowchartParentBegin("Flats");
-       for (var i = 0; i < filtered_flats.allfilesarr.length; i++) {
-             var filterFiles = filtered_flats.allfilesarr[i].files;
-             var filterName = filtered_flats.allfilesarr[i].filter;
-             this.flowchart.flowchartChildBegin(filterName);
-             if (this.par.flat_master_files.val) {
-                  // We have an array of master flat files, we match them by resolution
-                   this.util.addProcessingStep("calibrateEngine use existing " + filterName + " master flat files");
-                   masterflatPath[i] = filterFiles.length == 1 ? filterFiles[0].name : filterFiles.map(function(f) { return f.name; });
-             } else if (filterFiles.length == 1) {
-                   this.util.addProcessingStep("calibrateEngine use existing " + filterName + " master flat " + filterFiles[0].name);
-                   masterflatPath[i] = filterFiles[0].name;
-             } else if (filterFiles.length > 0) {
-                   // calibrate flats for each filter with master bias and/or master dark and/or dark flat if available
-                   this.util.addProcessingStep("calibrateEngine calibrate " + filterName + " flats using " + filterFiles.length + " files, " + filterFiles[0].name);
-                   var flatcalimages = this.fileNamesToEnabledPathFromFilearr(filterFiles);
-                   console.writeln("flatcalimages[0] " + flatcalimages[0][1]);
-                   var selectedMasterDarkForFlat = null;
-                   if (this.par.use_darks_on_flat_calibrate.val) {
-                        var flatExptime = filterFiles[0].exptime;
-                        selectedMasterDarkForFlat = this.selectMasterDarkForExptime(masterdarkPath, flatExptime);
-                        if (selectedMasterDarkForFlat != null) {
-                              this.util.addProcessingStep("calibrateEngine selected master dark for flat exptime " + flatExptime + "s: " + selectedMasterDarkForFlat);
-                        }
-                  }
-                   var flatcalFileNames = this.runCalibrateFlats(flatcalimages, masterbiasPath, selectedMasterDarkForFlat, masterflatdarkPath ? masterflatdarkPath[i] : null, filterName);
-                   console.writeln("flatcalFileNames[0] " + flatcalFileNames[0]);
- 
-                   // integrate flats to generate master flat for each filter
-                   var flatimages = this.filesForImageIntegration(flatcalFileNames);
-                   console.writeln("flatimages[0] " + flatimages[0][1]);
-                   let masterflatid = this.runImageIntegrationFlats(flatimages, this.ppar.win_prefix + "AutoMasterFlat_" + filterName, filterName);
-                   console.writeln("masterflatid " + masterflatid);
-                   this.setImagetypKeyword(this.util.findWindow(masterflatid), "Master flat");
-                   this.setFilterKeyword(this.util.findWindow(masterflatid), filterFiles[0].filter);
-                   masterflatPath[i] = this.saveMasterWindow(this.global.outputRootDir, masterflatid);
-                   this.guiUpdatePreviewId(masterflatid);
-             } else {
-                   masterflatPath[i] = null;
-             }
-             this.flowchart.flowchartChildEnd(filterName);
-       }
-       this.flowchart.flowchartParentEnd("Flats");
-
-       if (this.par.generate_masters_only.val) {
-             this.util.addProcessingStep("calibrateEngine, master calibration files generated");
-             this.util.runGarbageCollection();
-             return [ [], '' ];
-       }
-
-       /*
-        * Calibrate light files
-        *
-        * Calibrate light files with master dark and master flat. 
-        * Optionally, we can use also master bias for calibration.
-        * We calibrate light files for each filter separately using the corresponding master flat for the filter
-        * and correcponding exposure time for the master dark.
-        */
-       this.util.addProcessingStepAndStatusInfo("Image calibration, calibrate light images");
-       var calibratedLightFileNames = [];
-       this.flowchart.flowchartParentBegin("Calibrate lights");
-       for (var i = 0; i < filtered_lights.allfilesarr.length; i++) {
-             var filterFiles = filtered_lights.allfilesarr[i].files;
-             var filterName = filtered_lights.allfilesarr[i].filter;
-             this.flowchart.flowchartChildBegin(filterName);
-             if (filterFiles.length > 0) {
-                   // Sub-group lights by exposure time so each group gets the
-                   // correct master dark when multiple exposure times are present.
-                   var lightExptimeGroups = this.groupLightsByExposureTime(filterFiles);
-                   for (var lg = 0; lg < lightExptimeGroups.length; lg++) {
-                         var lgFiles   = lightExptimeGroups[lg].files;
-                         var lgExptime = lightExptimeGroups[lg].exptime;
-
-                         // calibrate light frames with master bias, master dark and master flat
-                         // optionally master dark can be left out
-                         let fileProcessedStatus = this.getFileProcessedStatusCalibrated(this.fileNamesFromFilearr(lgFiles), '_c');
-                         if (fileProcessedStatus.unprocessed.length == 0) {
-                               var node = this.flowchart.flowchartOperation("ImageCalibration:lights");
-                               calibratedLightFileNames = calibratedLightFileNames.concat(fileProcessedStatus.processed);
-                               this.engine_end_process(node);
-                         } else {
-                               this.util.addProcessingStep("calibrateEngine calibrate " + filterName + " lights for " + fileProcessedStatus.unprocessed.length + " files at " + lgExptime + "s");
-                               let lightcalimages = this.fileNamesToEnabledPath(fileProcessedStatus.unprocessed);
-
-                               var selectedMasterDark = this.selectMasterDarkForExptime(masterdarkPath, lgExptime);
-                               if (selectedMasterDark != null) {
-                                    this.util.addProcessingStep("calibrateEngine selected master dark for light exptime " + lgExptime + "s: " + selectedMasterDark);
-                               }
-                               let lightcalFileNames = this.runCalibrateLights(lightcalimages, masterbiasPath, selectedMasterDark, masterflatPath[i], filterName, lgExptime);
-
-                               calibratedLightFileNames = calibratedLightFileNames.concat(lightcalFileNames, fileProcessedStatus.processed);
-                         }
-                   }
-             }
-             this.flowchart.flowchartChildEnd(filterName);
-       }
-       this.flowchart.flowchartParentEnd("Calibrate lights");
-
-       // We now have calibrated light images
-       // We now proceed with cosmetic correction and
-       // after that debayering in case of OSC/RAW files
-
-       console.writeln("calibrateEngine, return calibrated images, calibratedLightFileNames[0] " + calibratedLightFileNames[0]);
-
-       this.util.runGarbageCollection();
-
-       return [ calibratedLightFileNames, '_c' ];
-}
  
 /***************************************************************************
  * 
@@ -18850,8 +17898,8 @@ engineInit()
             this.util.setDefaultDirs();
       }
       this.initSPCCvalues();
-      this.creating_mask = false;
-      this.skip_process_value_save = false;
+      this.global.creating_mask = false;
+      this.global.skip_process_value_save = false;
 }
 
 // Check possible conflicting settings before stating the processing
@@ -19049,7 +18097,7 @@ autointegrateProcessingEngine(parent, auto_continue, autocontinue_narrowband, tx
        var RGB_processed_HT_id = null;
        var LRGB_Combined = null;
  
-       this.executed_processes = [];
+       this.util.executed_processes = [];
        this.is_rgb_files = false;
        this.is_narrowband_files = false;
        this.is_color_files = false;
@@ -19761,10 +18809,10 @@ autointegrateProcessingEngine(parent, auto_continue, autocontinue_narrowband, tx
             console.noteln("Console output is written into file " + this.logfname);
       }
 
-      if (this.executed_processes.length > 0 && this.par.create_process_icons.val) {
+      if (this.util.executed_processes.length > 0 && this.par.create_process_icons.val) {
             let filename = this.util.ensure_win_prefix("ExecutedProcesses.xpsm");
             console.writeln("Write executed processes as process icons");
-            this.writeExecutedProcessesToXPSM(this.util.ensurePathEndSlash(this.global.outputRootDir) + filename);
+            this.util.writeExecutedProcessesToXPSM(this.util.ensurePathEndSlash(this.global.outputRootDir) + filename);
       }
       this.global.is_processing = this.global.processing_state.none;
 
@@ -19781,27 +18829,15 @@ printProcessValues(obj, txt = "")
       }
 }
 
-saveProcessValues(obj, txt = "")
+saveProcessValues(obj, txt = "", imageId = null)
 {
-      if (!this.global.get_flowchart_data
-          && this.global.is_processing != this.global.processing_state.none
-          && !this.creating_mask
-          && !this.skip_process_value_save)
-      {
-            this.executed_processes.push({ obj: obj, txt: txt });
-      }
+      this.util.addExecutedProcess(obj, txt, imageId);
 }
 
-printAndSaveProcessValues(obj, txt = "")
+printAndSaveProcessValues(obj, txt = "", imageId = null)
 {
       this.printProcessValues(obj, txt);
-      this.saveProcessValues(obj, txt);
-}
-
-printAndSaveProcessValues(obj, txt = "")
-{
-      this.printProcessValues(obj, txt);
-      this.saveProcessValues(obj, txt);
+      this.saveProcessValues(obj, txt, imageId);
 }
 
 printProcessDefaultValues(name, obj)
@@ -19810,55 +18846,6 @@ printProcessDefaultValues(name, obj)
       console.writeln(obj.toSource());
 }
 
-writeExecutedProcessesToXPSM(filename)
-{
-      console.writeln("Write executed processes to XPSM file " + filename);
-      var file = new File();
-      file.createForWriting(filename);
-
-      file.outTextLn('<?xml version="1.0" encoding="UTF-8"?>');
-      file.outTextLn('<!--');
-      file.outTextLn('********************************************************************');
-      file.outTextLn('PixInsight XML Process Serialization Module - XPSM 1.0');
-      file.outTextLn('AutoIntegrate processings steps');
-      file.outTextLn('********************************************************************');
-      file.outTextLn('Generated on ' + (new Date()).toString());
-      file.outTextLn('PixInsight version ' + this.global.pixinsight_version_str);
-      file.outTextLn('AutoIntegrate version ' + this.global.autointegrate_version);
-      file.outTextLn('********************************************************************');
-      file.outTextLn('-->');
-
-      file.outText('<xpsm version=\"1.0\"');
-      file.outText(' mlns="http://www.pixinsight.com/xpsm"');
-      file.outText(' xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"');
-      file.outText(' xsi:schemaLocation="http://www.pixinsight.com/xpsm');
-      file.outTextLn(' http://pixinsight.com/xpsm/xpsm-1.0.xsd">');
-
-      // Write process instances
-      for (var i = 0; i < this.executed_processes.length; i++) {
-            var src = this.executed_processes[i].obj.toSource("XPSM 1.0");
-            var processId = this.executed_processes[i].obj.processId();
-            file.outTextLn('<!-- ' + processId + (i+1) + ' -->');
-            src = src.replace(processId + "_instance", "ai_" + processId + "_instance" + (i+1));
-            src = src.replace(/Integration_([LRGBSHOC])_map/g, "Integration_$1");
-            src = src.replace(/Integration_RGB_map/g, "Integration_RGB");
-            file.outTextLn(src);
-      }
-
-      // write icons
-      for (var i = 0; i < this.executed_processes.length; i++) {
-            var processId = this.executed_processes[i].obj.processId();
-            var txt = this.executed_processes[i].txt;
-            if (txt != "" && txt != null) {
-                  txt = '_' + txt;
-            }
-            file.outTextLn('<icon id="ai_' + processId + txt + '_' + (i+1) + '"' +
-                           ' instance="ai_' + processId + '_instance' + (i+1) + '"' +
-                           ' xpos="' + (this.global.screen_width - 600) + '" ypos="' + (5 + 32 * (i+1)) + '" workspace="Workspace01"/>');
-      }
-      file.outTextLn('</xpsm>\n');
-      file.close();
-}
 
 getProcessDefaultValues()
 {
@@ -19969,7 +18956,7 @@ this.openDirectoryFiles = openDirectoryFiles;
 this.getFilterFiles = getFilterFiles;
 this.getFilterHigh = getFilterHigh;
 this.getImagetypFiles = getImagetypFiles;
-this.getExptimeFromFile = getExptimeFromFile;
+this.getExptimeFromFile = (filePath) => this.calibrate.getExptimeFromFile(filePath);
 this.runResample = runResample;
 
 this.writeProcessingStepsAndEndLog = writeProcessingStepsAndEndLog;
